Fix: Fixed issue #5, market_get_items now uses Pipelining.
Speed increase is dramatic, however, Broken items that do not have the
quality specified in their key name will now be "unknown" and will act
like brand new items.

The original issue that the API created was solved and the code to
perform an alternate lookup was just a workaround, nearly all items
should now have been moved to the correct keys as they've been traded.