Hopefully fix multiple stacks of items causing exist test false positive.
If multiple stacks of items were sent like two lots of silver,
while iterating the list of Things, the hash would only get added as a
dictionary key once, but the exist test would be queued in Redis pipe.

When the zip was performed, the non-existent item would be potentially
paired with a "True" result even though it doesn't exist, causing a
cascading issue.

E.g.

List of goods: [A, A, B, C, D, D, E],
Assume goods A, C, and D exist but B and E do not.
This would result in a list of dictionary keys to iterate like this:
[A, B, C, D, E] but the results list is: [T, T, F, T, T, T, F]
When zipped it does this: [A: T, B: T, C: F, D: T, E: T, F: T]

Notice how B and F are now both True but C is False when it should be
True because the duplicate where compressed into a single key.


(cherry picked from commit 72929116f5074518e5ab49a9dc7bf035ba3614c4)